* 'master' of https://github.com/vyloy/skywire:
  add version functionality
  [doc] Deleting chinise link
  Added spanish document link
  Fixed docker shortcut
  spanish version of readme file added
  Fixing Chinise link

+++ cmd/manager/main.go
@@ -2,14 +2,15 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	log "github.com/sirupsen/logrus"
 	"github.com/skycoin/net/skycoin-messenger/factory"
 	"github.com/skycoin/net/skycoin-messenger/monitor"
 	"github.com/skycoin/skycoin/src/util/file"
+	"github.com/skycoin/skywire/manager"
 	"os"
 	"os/signal"
 	"path/filepath"
-	"github.com/skycoin/skywire/cmd"
 )
 
 var (
@@ -18,8 +19,7 @@ var (
 	webPort  string
 	seedPath string
 
-	code    string
-	version string
+	version bool
 )
 
 func parseFlags() {
@@ -28,11 +28,16 @@ func parseFlags() {
 	flag.StringVar(&webPort, "web-port", ":8000", "monitor web page port")
 	flag.StringVar(&address, "address", ":5998", "address to listen on")
 	flag.StringVar(&seedPath, "seed-path", filepath.Join(file.UserHome(), ".skywire", "discovery", "keys.json"), "path to save seed info")
+	flag.BoolVar(&version, "v", false, "print current version")
 	flag.Parse()
 }
 
 func main() {
 	parseFlags()
+	if version {
+		fmt.Println(manager.Version)
+		return
+	}
 
 	osSignal := make(chan os.Signal, 1)
 	signal.Notify(osSignal, os.Interrupt, os.Kill)
@@ -41,13 +46,14 @@ func main() {
 	defer f.Close()
 	f.SetDefaultSeedConfigPath(seedPath)
 	f.SetLoggerLevel(factory.DebugLevel)
+	f.SetAppVersion(manager.Version)
 	err := f.Listen(address)
 	log.Debugf("listen on %s", address)
 	if err != nil {
 		log.Error(err)
 		os.Exit(1)
 	}
-	m := monitor.New(f, address, webPort, cmd.MANAGER_TAG, cmd.MANAGER_VERSION)
+	m := monitor.New(f, address, webPort, manager.Tag, manager.Version)
 	m.Start(webDir)
 	defer m.Close()
 	select {
